# openapi_mf1.yaml
openapi: 3.1.0
info:
  title: E-Receipt IT API
  description: E-Receipt IT API
  version: 1.0.0
paths:
...
  /mf1/point-of-sales:
    get:
      tags:
      - Point of Sale
      summary: Read Point Of Sales
      description: Retrieve PEMs.
      operationId: read_point_of_sales_mf1_point_of_sales_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/PEMStatus'
          - type: 'null'
          title: Status
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 30
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page__T_Customized_PointOfSaleOutput_'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel403Forbidden'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel404NotFound'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /mf1/point-of-sales/{serial_number}:
    get:
      tags:
      - Point of Sale
      summary: Read Point Of Sale
      operationId: read_point_of_sale_mf1_point_of_sales__serial_number__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: serial_number
        in: path
        required: true
        schema:
          type: string
          description: The Point of Sale's serial number
          title: Serial Number
        description: The Point of Sale's serial number
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointOfSaleDetailedOutput'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel403Forbidden'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel404NotFound'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /mf1/point-of-sales/close:
    post:
      tags:
      - Point of Sale
      summary: Close Journal
      operationId: close_journal_mf1_point_of_sales_close_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel403Forbidden'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel404NotFound'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /mf1/point-of-sales/{serial_number}/activation:
    post:
      tags:
      - Point of Sale
      summary: Post Activation
      description: Trigger the activation process of a Point of Sale by requesting
        a certificate to the Italian Tax Agency
      operationId: post_activation_mf1_point_of_sales__serial_number__activation_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: serial_number
        in: path
        required: true
        schema:
          type: string
          description: The Point of Sale's serial number
          title: Serial Number
        description: The Point of Sale's serial number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivationRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Post Activation Mf1 Point Of Sales  Serial Number  Activation
                  Post
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel403Forbidden'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel404NotFound'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /mf1/point-of-sales/{serial_number}/inactivity:
    post:
      tags:
      - Point of Sale
      summary: Create Inactivity Period
      description: Create a new inactivity period
      operationId: create_inactivity_period_mf1_point_of_sales__serial_number__inactivity_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: serial_number
        in: path
        required: true
        schema:
          type: string
          description: The Point of Sale serial number
          title: Serial Number
        description: The Point of Sale serial number
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Create Inactivity Period Mf1 Point Of Sales  Serial
                  Number  Inactivity Post
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel403Forbidden'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel404NotFound'
          description: Not Found
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel401Unauthorized'
          description: Unauthorized
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel400BadRequest'
          description: Bad Request
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /mf1/point-of-sales/{serial_number}/status/offline:
    post:
      tags:
      - Point of Sale
      summary: Post Offline
      description: Change the state of the Point of Sale to 'offline'
      operationId: post_offline_mf1_point_of_sales__serial_number__status_offline_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: serial_number
        in: path
        required: true
        schema:
          type: string
          description: The Point of Sale serial number
          title: Serial Number
        description: The Point of Sale serial number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PEMStatusOfflineRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Post Offline Mf1 Point Of Sales  Serial Number  Status
                  Offline Post
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel403Forbidden'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel404NotFound'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ActivationRequest:
      properties:
        registration_key:
          type: string
          title: Registration Key
      type: object
      required:
      - registration_key
      title: ActivationRequest
    Address:
      properties:
        street_address:
          type: string
          title: Street Address
          description: The street address associated to the PEM
        street_number:
          type: string
          title: Street Number
          description: The street number associated to the PEM
        zip_code:
          type: string
          maxLength: 5
          minLength: 5
          title: Zip Code
          description: The zip code associated to the PEM
        city:
          type: string
          title: City
          description: The city associated to the PEM
        province:
          type: string
          maxLength: 2
          minLength: 2
          title: Province
          description: The province associated to the PEM
      type: object
      required:
      - street_address
      - street_number
      - zip_code
      - city
      - province
      title: Address
...
    ErrorModel400BadRequest:
      properties:
        type:
          type: string
          title: Type
          default: /errors/400
        title:
          type: string
          title: Title
          default: Bad Request
        status:
          type: integer
          title: Status
          default: 400
        detail:
          type: string
          title: Detail
          default: A human-readable explanation specific to this occurrence of the
            problem
        instance:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance
          default: A URI reference that identifies the specific occurrence of the
            problem
      type: object
      title: ErrorModel400BadRequest
    ErrorModel401Unauthorized:
      properties:
        type:
          type: string
          title: Type
          default: /errors/401
        title:
          type: string
          title: Title
          default: Could not validate credentials
        status:
          type: integer
          title: Status
          default: 401
        detail:
          type: string
          title: Detail
          default: A human-readable explanation specific to this occurrence of the
            problem
        instance:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance
          default: A URI reference that identifies the specific occurrence of the
            problem
      type: object
      title: ErrorModel401Unauthorized
    ErrorModel403Forbidden:
      properties:
        type:
          type: string
          title: Type
          default: /errors/403
        title:
          type: string
          title: Title
          default: Forbidden
        status:
          type: integer
          title: Status
          default: 403
        detail:
          type: string
          title: Detail
          default: A human-readable explanation specific to this occurrence of the
            problem
        instance:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance
          default: A URI reference that identifies the specific occurrence of the
            problem
      type: object
      title: ErrorModel403Forbidden
    ErrorModel404NotFound:
      properties:
        type:
          type: string
          title: Type
          default: /errors/404
        title:
          type: string
          title: Title
          default: Not Found
        status:
          type: integer
          title: Status
          default: 404
        detail:
          type: string
          title: Detail
          default: A human-readable explanation specific to this occurrence of the
            problem
        instance:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance
          default: A URI reference that identifies the specific occurrence of the
            problem
      type: object
      title: ErrorModel404NotFound
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PEMStatus:
      type: string
      enum:
      - NEW
      - REGISTERED
      - ACTIVATED
      - ONLINE
      - OFFLINE
      - DISCARDED
      title: PEMStatus
    PEMStatusOfflineRequest:
      properties:
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        reason:
          type: string
          title: Reason
      type: object
      required:
      - timestamp
      - reason
      title: PEMStatusOfflineRequest
...
    Page__T_Customized_PointOfSaleOutput_:
      properties:
        members:
          items:
            $ref: '#/components/schemas/PointOfSaleOutput'
          type: array
          title: Members
        total:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Total
        page:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Page
        size:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Size
        pages:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Pages
      type: object
      required:
      - members
      - page
      - size
      title: Page[~T]Customized[PointOfSaleOutput]
...
    PointOfSaleDetailedOutput:
      properties:
        serial_number:
          type: string
          title: Serial Number
        status:
          $ref: '#/components/schemas/PEMStatus'
        address:
          $ref: '#/components/schemas/Address'
        registration_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Registration Key
      type: object
      required:
      - serial_number
      - status
      - address
      - registration_key
      title: PointOfSaleDetailedOutput
    PointOfSaleOutput:
      properties:
        serial_number:
          type: string
          title: Serial Number
        status:
          $ref: '#/components/schemas/PEMStatus'
        address:
          $ref: '#/components/schemas/Address'
      type: object
      required:
      - serial_number
      - status
      - address
      title: PointOfSaleOutput
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /mf1/login